refactor async url analyze prebuilt tests
