--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_business_card_from_url_async.test_authentication_bad_key.yaml
@@ -1,29 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"urlSource": "blob_sas_url"}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '232'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.1 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-businessCard:analyze?stringIndexType=unicodeCodePoint&api-version=2021-07-30-preview
-  response:
-    body:
-      string: '{"error": {"code": "401", "message": "Access denied due to invalid
-        subscription key or wrong API endpoint. Make sure to provide a valid key for
-        an active subscription and use a correct regional API endpoint for your resource."}}'
-    headers:
-      apim-request-id: 7e961977-912e-4c97-8eaa-41474d606c72
-      content-length: '224'
-      content-type: application/json
-      date: Tue, 14 Sep 2021 08:46:42 GMT
-    status:
-      code: 401
-      message: PermissionDenied
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-businessCard:analyze?stringIndexType=unicodeCodePoint&api-version=2021-07-30-preview
-version: 1
