--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_business_card_from_url_async.test_business_card_bad_url.yaml
@@ -1,33 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"urlSource": "https://badurl.jpg"}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '35'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.1 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-businessCard:analyze?stringIndexType=unicodeCodePoint&api-version=2021-07-30-preview
-  response:
-    body:
-      string: '{"error": {"code": "InvalidRequest", "message": "Invalid request.",
-        "innererror": {"code": "InvalidContent", "message": "The file is corrupted
-        or format is unsupported. Refer to documentation for the list of supported
-        formats."}}}'
-    headers:
-      apim-request-id: 6cbfa980-5e96-4b07-8a5d-41ffafc4c5f5
-      content-type: application/json; charset=utf-8
-      date: Tue, 14 Sep 2021 08:47:12 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '102'
-    status:
-      code: 400
-      message: Bad Request
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-businessCard:analyze?stringIndexType=unicodeCodePoint&api-version=2021-07-30-preview
-version: 1
