--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_business_card_from_url_async.test_business_card_locale_error.yaml
@@ -1,32 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"urlSource": "blob_sas_url"}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '232'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.1 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-businessCard:analyze?locale=not%20a%20locale&stringIndexType=unicodeCodePoint&api-version=2021-07-30-preview
-  response:
-    body:
-      string: '{"error": {"code": "InvalidArgument", "message": "Invalid argument.",
-        "innererror": {"code": "InvalidParameter", "message": "The parameter Locale
-        is invalid: The language code is invalid or not supported."}}}'
-    headers:
-      apim-request-id: 2a61b0fe-db31-45f9-a4e4-db723e5cba8f
-      content-type: application/json; charset=utf-8
-      date: Tue, 14 Sep 2021 08:45:20 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '871'
-    status:
-      code: 400
-      message: Bad Request
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-businessCard:analyze?locale=not%20a%20locale&stringIndexType=unicodeCodePoint&api-version=2021-07-30-preview
-version: 1
