--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_business_card_from_url_async.test_business_card_url_pass_stream.yaml
@@ -1,32 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"urlSource": "<_io.BufferedReader name=\''c:\\\\\\\\Users\\\\\\\\caperal\\\\\\\\Documents\\\\\\\\Python\\\\\\\\azure-sdk-for-python-pr\\\\\\\\sdk\\\\\\\\formrecognizer\\\\\\\\azure-ai-formrecognizer\\\\\\\\tests\\\\\\\\sample_forms\\\\\\\\business_cards\\\\\\\\business-card-english.png\''>"}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '241'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.1 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-businessCard:analyze?stringIndexType=unicodeCodePoint&api-version=2021-07-30-preview
-  response:
-    body:
-      string: '{"error": {"code": "InvalidArgument", "message": "Invalid argument.",
-        "innererror": {"code": "InvalidContentSourceFormat", "message": "Invalid content
-        source: Url must be http or https."}}}'
-    headers:
-      apim-request-id: 62c17083-5ac6-4bae-a80f-29deb6eb7994
-      content-type: application/json; charset=utf-8
-      date: Tue, 14 Sep 2021 08:47:17 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '14'
-    status:
-      code: 400
-      message: Bad Request
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-businessCard:analyze?stringIndexType=unicodeCodePoint&api-version=2021-07-30-preview
-version: 1
