--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_receipt_from_url_async.test_receipt_bad_url.yaml
@@ -1,33 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"urlSource": "https://badurl.jpg"}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '35'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-receipt:analyze?stringIndexType=unicodeCodePoint&api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"error": {"code": "InvalidRequest", "message": "Invalid request.",
-        "innererror": {"code": "InvalidContent", "message": "The file is corrupted
-        or format is unsupported. Refer to documentation for the list of supported
-        formats."}}}'
-    headers:
-      apim-request-id: 2622b562-e74e-4a5f-9b64-b439a8708e16
-      content-type: application/json; charset=utf-8
-      date: Tue, 14 Sep 2021 00:28:19 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '131'
-    status:
-      code: 400
-      message: Bad Request
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-receipt:analyze?stringIndexType=unicodeCodePoint&api-version=2021-09-30-preview
-version: 1
