--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_receipt_from_url_async.test_receipt_locale_error.yaml
@@ -1,32 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"urlSource": "blob_sas_url"}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '224'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-receipt:analyze?locale=not%20a%20locale&stringIndexType=unicodeCodePoint&api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"error": {"code": "InvalidArgument", "message": "Invalid argument.",
-        "innererror": {"code": "InvalidParameter", "message": "The parameter Locale
-        is invalid: The language code is invalid or not supported."}}}'
-    headers:
-      apim-request-id: e6c4d122-9c19-4bfb-9d69-f5914cfc29b4
-      content-type: application/json; charset=utf-8
-      date: Thu, 30 Sep 2021 17:06:34 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '581'
-    status:
-      code: 400
-      message: Bad Request
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-receipt:analyze?locale=not%20a%20locale&stringIndexType=unicodeCodePoint&api-version=2021-09-30-preview
-version: 1
