--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_receipt_from_url_async.test_receipt_url_auth_bad_key.yaml
@@ -1,32 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"urlSource": "blob_sas_url"}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '224'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-receipt:analyze?stringIndexType=unicodeCodePoint&api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"error": {"code": "401", "message": "Access denied due to invalid
-        subscription key or wrong API endpoint. Make sure to provide a valid key for
-        an active subscription and use a correct regional API endpoint for your resource."}}'
-    headers:
-      apim-request-id: bb36ce70-73f6-43a6-a7a6-a380bafc6ff4
-      content-length: '225'
-      content-type: application/json
-      date: Tue, 14 Sep 2021 00:28:52 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      www-authenticate: AzureApiManagementKey realm="https://region.api.cognitive.microsoft.com/formrecognizer",name="Ocp-Apim-Subscription-Key",type="header"
-      x-content-type-options: nosniff
-    status:
-      code: 401
-      message: Access Denied
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-receipt:analyze?stringIndexType=unicodeCodePoint&api-version=2021-09-30-preview
-version: 1
