--- sdk/formrecognizer/azure-ai-formrecognizer/tests/recordings/test_receipt_from_url_async.test_receipt_url_pass_stream.yaml
@@ -1,32 +1,0 @@
-interactions:
-- request:
-    body: 'b''{"urlSource": "b\''\\\\x89PNG\''"}'''
-    headers:
-      Accept:
-      - application/json
-      Content-Length:
-      - '28'
-      Content-Type:
-      - application/json
-      User-Agent:
-      - azsdk-python-ai-formrecognizer/3.2.0b1 Python/3.9.0 (Windows-10-10.0.19041-SP0)
-    method: POST
-    uri: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-receipt:analyze?stringIndexType=unicodeCodePoint&api-version=2021-09-30-preview
-  response:
-    body:
-      string: '{"error": {"code": "InvalidArgument", "message": "Invalid argument.",
-        "innererror": {"code": "InvalidContentSourceFormat", "message": "Invalid content
-        source: Url must be http or https."}}}'
-    headers:
-      apim-request-id: d93f5a6e-f030-4f15-9760-0cfed562b796
-      content-type: application/json; charset=utf-8
-      date: Tue, 14 Sep 2021 00:28:59 GMT
-      strict-transport-security: max-age=31536000; includeSubDomains; preload
-      transfer-encoding: chunked
-      x-content-type-options: nosniff
-      x-envoy-upstream-service-time: '11'
-    status:
-      code: 400
-      message: Bad Request
-    url: https://region.api.cognitive.microsoft.com/formrecognizer/documentModels/prebuilt-receipt:analyze?stringIndexType=unicodeCodePoint&api-version=2021-09-30-preview
-version: 1
