+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_identity_documents_from_url_async.py
@@ -11,12 +11,11 @@ from azure.core.credentials import AzureKeyCredential
 from azure.ai.formrecognizer._generated.v2_1.models import AnalyzeOperationResult
 from azure.ai.formrecognizer._response_handlers import prepare_prebuilt_models
 from azure.ai.formrecognizer import FormRecognizerApiVersion
-from azure.ai.formrecognizer.aio import FormRecognizerClient, DocumentAnalysisClient
+from azure.ai.formrecognizer.aio import FormRecognizerClient
 from asynctestcase import AsyncFormRecognizerTest
 from preparers import FormRecognizerPreparer
 from preparers import GlobalClientPreparer as _GlobalClientPreparer
 
-DocumentAnalysisClientPreparer = functools.partial(_GlobalClientPreparer, DocumentAnalysisClient)
 FormRecognizerClientPreparer = functools.partial(_GlobalClientPreparer, FormRecognizerClient)
 
 
@@ -70,49 +69,6 @@ class TestIdDocumentsFromUrlAsync(AsyncFormRecognizerTest):
 
         # Check page metadata
         self.assertFormPagesTransformCorrect(id_document.pages, read_results, page_results)
-
-    @FormRecognizerPreparer()
-    @DocumentAnalysisClientPreparer()
-    async def test_identity_document_jpg_passport(self, client):
-        async with client:
-            poller = await client.begin_analyze_document_from_url("prebuilt-idDocument", self.identity_document_url_jpg_passport)
-
-            result = await poller.result()
-            assert len(result.documents) == 1
-        
-            id_document = result.documents[0]
-            # check dict values
-
-            passport = id_document.fields.get("MachineReadableZone").value
-            assert passport["LastName"].value == "MARTIN"
-            assert passport["FirstName"].value == "SARAH"
-            assert passport["DocumentNumber"].value == "ZE000509"
-            assert passport["DateOfBirth"].value == date(1985,1,1)
-            assert passport["DateOfExpiration"].value == date(2023,1,14)
-            assert passport["Sex"].value == "F"
-            assert passport["CountryRegion"].value == "CAN"
-
-    @FormRecognizerPreparer()
-    @DocumentAnalysisClientPreparer()
-    async def test_identity_document_jpg(self, client):
-        async with client:
-            poller = await client.begin_analyze_document_from_url("prebuilt-idDocument", self.identity_document_url_jpg)
-
-            result = await poller.result()
-        assert len(result.documents) == 1
-        id_document = result.documents[0]
-
-        # check dict values
-        assert id_document.fields.get("LastName").value == "TALBOT"
-        assert id_document.fields.get("FirstName").value == "LIAM R."
-        assert id_document.fields.get("DocumentNumber").value == "WDLABCD456DG"
-        assert id_document.fields.get("DateOfBirth").value == date(1958,1,6)
-        assert id_document.fields.get("DateOfExpiration").value == date(2020,8,12)
-        assert id_document.fields.get("Sex").value == "M"
-        assert id_document.fields.get("Address").value == "123 STREET ADDRESS YOUR CITY WA 99999-1234"
-        assert id_document.fields.get("CountryRegion").value == "USA"
-        assert id_document.fields.get("Region").value == "Washington"
-
 
     @FormRecognizerPreparer()
     @FormRecognizerClientPreparer()
